feat: 그룹 페이지네이션 컴포넌트를 공통 컴포넌트로 분리
